services/Prosopography_01/requirements.txt
# Prosopography Tool Dependencies

# Web framework
streamlit>=1.28.0

# Database
psycopg2-binary>=2.9.0

# LLM
cohere>=4.0.0

# HTTP requests
requests>=2.28.0

# HTML parsing
beautifulsoup4>=4.12.0

# PDF parsing (optional)
PyPDF2>=3.0.0

# Environment variables
python-dotenv>=1.0.0
